Rename testutils package in OpenAPI backend to use name 'sample'
