+++ src/python/pants/backend/openapi/sample/petstore_spec.yaml
