+++ src/python/pants/backend/openapi/sample/resources.py
@@ -3,6 +3,6 @@
 
 import importlib.resources as pkg_resources
 
-from pants.backend.openapi import testutils
+from pants.backend.openapi import sample
 
-PETSTORE_SAMPLE_SPEC = pkg_resources.read_text(testutils.__name__, "petstore_spec.yaml")
+PETSTORE_SAMPLE_SPEC = pkg_resources.read_text(sample.__name__, "petstore_spec.yaml")
